Clean up routes
