+++ src/services/chat/tools.ts
@@ -1,19 +1,5 @@
 import { createMessage, updateMessage } from "../../db/operations";
-import { getTool } from "../../tools";
-import type { ConversationState, State } from "../../types/core";
 import logger from "../../utils/logger";
-import { calculateRequestPrice } from "../../x402/pricing";
-
-export type ToolResult =
-  | { ok: true; data?: unknown }
-  | { ok: false; error: string };
-
-export interface ToolExecutionContext {
-  state: State;
-  conversationState: ConversationState;
-  message: any;
-  files?: File[];
-}
 
 export interface MessageCreationParams {
   conversationId: string;
@@ -73,204 +59,6 @@ export async function createMessageRecord(
   } catch (err) {
     if (logger) logger.error({ err, isExternal }, "create_message_failed");
     return { success: false, error: "Failed to create message" };
-  }
-}
-
-/**
- * Execute file upload tool
- */
-export async function executeFileUpload(
-  context: ToolExecutionContext,
-): Promise<{ success: boolean; error?: string }> {
-  const { state, conversationState, message, files } = context;
-
-  if (!files || files.length === 0) {
-    return { success: true };
-  }
-
-  const fileUploadTool = getTool("FILE-UPLOAD");
-  if (!fileUploadTool) {
-    return { success: true }; // Not a critical failure
-  }
-
-  try {
-    if (logger) {
-      logger.info(
-        `Processing ${files.length} uploaded file(s) before planning`,
-      );
-    }
-    await fileUploadTool.execute({
-      state,
-      conversationState,
-      message,
-      files,
-    });
-    return { success: true };
-  } catch (err) {
-    if (logger) logger.error({ err }, "file_upload_failed");
-    return { success: false, error: "Failed to process uploaded files" };
-  }
-}
-
-/**
- * Execute planning tool
- */
-export async function executePlanning(context: ToolExecutionContext): Promise<{
-  success: boolean;
-  result?: { providers: string[]; actions: string[] };
-  error?: string;
-}> {
-  const { state, conversationState, message } = context;
-
-  const planningTool = getTool("PLANNING");
-  if (!planningTool) {
-    return { success: false, error: "Planning tool not found" };
-  }
-
-  try {
-    const planningResult = await planningTool.execute({
-      state,
-      conversationState,
-      message,
-    });
-
-    if (logger) {
-      logger.info(
-        {
-          providers: planningResult.providers,
-          action: planningResult.actions[0],
-        },
-        "Executing plan",
-      );
-    }
-
-    return { success: true, result: planningResult };
-  } catch (err) {
-    if (logger) logger.error({ err }, "planning_tool_failed");
-    return { success: false, error: "Planning tool execution failed" };
-  }
-}
-
-/**
- * Execute provider tools in parallel
- */
-export async function executeProviderTools(
-  providers: string[],
-  context: ToolExecutionContext,
-): Promise<void> {
-  const { state, conversationState, message } = context;
-
-  await Promise.all(
-    (providers ?? []).map(async (provider) => {
-      // Skip FILE-UPLOAD since we already processed it
-      if (provider === "FILE-UPLOAD") {
-        return { provider, result: { ok: true, data: {} } };
-      }
-
-      const tool = getTool(provider);
-      if (!tool) {
-        if (logger) logger.warn({ provider }, "provider_tool_missing");
-        const res: ToolResult = {
-          ok: false,
-          error: `Tool not found for provider: ${provider}`,
-        };
-        return { provider, result: res };
-      }
-
-      try {
-        // Initialize estimatedCostsUSD object if not exists
-        if (!state.values.estimatedCostsUSD) {
-          state.values.estimatedCostsUSD = {};
-        }
-
-        // Track cost for this provider before execution
-        const providerCost = calculateRequestPrice([provider]);
-        state.values.estimatedCostsUSD[provider] = parseFloat(providerCost);
-
-        const data = await tool.execute({
-          state,
-          conversationState,
-          message,
-        });
-        const res: ToolResult = { ok: true, data };
-        return { provider, result: res };
-      } catch (err) {
-        if (logger) logger.error({ provider, err }, "provider_tool_failed");
-        const res: ToolResult = {
-          ok: false,
-          error: `Tool execution failed for provider: ${provider}`,
-        };
-        return { provider, result: res };
-      }
-    }),
-  );
-}
-
-/**
- * Execute primary action tool (REPLY or HYPOTHESIS)
- */
-export async function executeActionTool(
-  action: string,
-  context: ToolExecutionContext,
-): Promise<{ success: boolean; result?: any; error?: string }> {
-  const { state, conversationState, message } = context;
-
-  const actionTool = getTool(action);
-  if (!actionTool) {
-    return { success: false, error: `Action tool not found: ${action}` };
-  }
-
-  // Track cost for action tool
-  if (!state.values.estimatedCostsUSD) {
-    state.values.estimatedCostsUSD = {};
-  }
-  const actionCost = calculateRequestPrice([action]);
-  state.values.estimatedCostsUSD[action] = parseFloat(actionCost);
-
-  try {
-    const result = await actionTool.execute({
-      state,
-      conversationState,
-      message,
-    });
-    return { success: true, result };
-  } catch (err) {
-    if (logger) logger.error({ action, err }, "action_tool_failed");
-    return { success: false, error: `Action tool execution failed: ${action}` };
-  }
-}
-
-/**
- * Execute reflection tool (optional)
- */
-export async function executeReflection(
-  context: ToolExecutionContext,
-): Promise<void> {
-  const { state, conversationState, message } = context;
-
-  const reflectionTool = getTool("REFLECTION");
-  if (!reflectionTool) {
-    if (logger) logger.warn("REFLECTION tool not found");
-    return;
-  }
-
-  try {
-    // Track cost for REFLECTION tool
-    if (!state.values.estimatedCostsUSD) {
-      state.values.estimatedCostsUSD = {};
-    }
-    const reflectionCost = calculateRequestPrice(["REFLECTION"]);
-    state.values.estimatedCostsUSD["REFLECTION"] = parseFloat(reflectionCost);
-
-    await reflectionTool.execute({
-      state,
-      conversationState,
-      message,
-    });
-    if (logger) logger.info("REFLECTION executed successfully");
-  } catch (err) {
-    if (logger) logger.error({ err }, "reflection_tool_failed");
-    // Don't fail the request if REFLECTION fails
   }
 }
